requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
google-auth==2.29.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
mysqlclient==2.2.4
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
requests==2.32.3
rsa==4.9
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
stripe==9.10.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.4.post2
